apps/api/pyproject.toml
[project]
name = "garag-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.21",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "redis>=5.0.0",
  "rq>=1.16.0",
  "boto3>=1.34.0",
  "opensearch-py>=2.5.0",
]

[tool.uvicorn]
factory = false